requirements.txt
flask==2.3.3
flask-login==0.6.2
werkzeug==2.3.7
flask-sqlalchemy==3.0.5
python-dotenv==1.0.0
cryptography==41.0.3
flask-mail==0.9.1
flask-cors==4.0.0
